Convert mocha to jest in testAuthenticatePassword.js
